Test can now be compiled and run
Rearranged test source such that test can be
compiled and run.  Note that the project to be
tested must also be installed on the emulator in
order for the installed test project to run.

Currently there is one test for the
CalFitWaveActivity class, which fails, as the
class is not yet written yet.