Decision point: Recipes as Services?
The current buildfile allows compilation, but no
generation of a signed package.  May be due to the
lack of definitions in the AndroidManifest.xml.
Two options come to mind: Package the dex and
description.xml into their own signed jar, or
implement recipes as services in android packages
with description.xml in the assets.