WaveRecipeTest#testRecipeOneAlgorithmService
Test only passes if Recipe apk is installed as
a usual android package on the device.  It cannot
seem to be found when it is loaded through the
path class loader.  It seems we need to either
reload the Intent resolver, or load the apk in a
different way.  It may be easier to return to the
non-service recipes and invoke with reflection.