Design simplification and refactor (partial)
WaveSensors and WaveSensorDescriptions have been
simplified in that their channel can no longer
specify units, units are only specified at the
sensor itself. This simplifies the process of
matching available sensors with recipes
significantly. Currently, the ability to specify
sampling rates and precisions per-channel remains.

I updated my SDK while in the process of these
changes, and while the project remains at API
level 4, this seems to have introduced an issue
with WaveRecipeAlgorithmShadow, a protected class
in the WaveRecip.java file.  A commit is created
now so that this bug can be traced and isolated.
Otherwise all previous tests which passed appear
to still pass.