Fix for bug introduced in commit 861f4a7dee979174d70b609f004d33dd7a7d56e8
That commit refactored some of the WaveService,
introducing an issue that did not show up until
later.  It seems that methods in the
mPublicBinder IWaveServicePublic.Stub cannot call
methods outside of that stub, possibly only when
such methods return null.

This caused a NullPointerException from
android.os.Parcel.readException(Parcel.java:1253)
when the IPC returns.