Added precisionHint to WaveSensor
previously we allowed a rate parameter in
WaveSensor.start, which gave a rate hint.  We now
request a precisionHint as well (primarily for use
with the Location sensor at this time).  We supply
this hint to the location sensor in meters, even
though the sensor reports in degrees latitude/
longitude (one cannot really specify accuracy in
fractions of a degree in order to preserve
similarity of units).  Unfortunately this means we
do need a special UI case for the
AndroidLocationSensor.