Major rework of sensor data dispatch
We have eliminated the WaveSensorListener
interface (and correspondingly, WaveSensorEvent),
and instead dispatch data through rate and
precision limiting inner classes.

Precision limiting and channel handling for the
AndroidLocationSensor are not yet implemented.

Most of the previous logging calls have been
removed, and test code is not yet updated to
match these changes.