Accelerometer performance fix!
Applied Looper use from AndroidLocationSensor to
AndroidHardwareSensor, based on a hunch.
AndroidWaveTesterClient now gets 30+Hz when
unrestricted.  This fix also applies to the
Magnetometer.

I speculate that somehow, the Android
SensorManager service was attaching to an
incorrect Looper, such as one for the UI, which
was causing it to block regularly.  Unabled to
tell from before/after traces.  The overhead of
Method.invoke seems to be minimal, escpecially
compared to my previous conjecture.